+++ CHANGELOG.md
@@ -727,3 +727,4 @@
 * Rename `?==` operator to `?=`
 * Remove `Conditional` module
 * Change operators' precedence, length of operators define precedence
+* Remove `leaf`, `left`, `right`, `both` in `Binary` module